The beginnings of the gui
Added the Single Play Pane for opening an audio file.  Getting the pane to move around was a bit tricky and needed a strange work around, but it works.

Now that the main function is working, hopefully more gui elements can be programmed in a bit faster.

The file dialog now forces focus so the user cannot make more dialogs.